Enum rubrail::SwipeState

pub enum SwipeState {
    Began,
    Changed,
    Ended,
    Cancelled,
    Failed,
    Unknown,
}

State of the current swipe gesture on an item

Variants
Began

Swipe gesture is newly detected (finger touched)

Changed

Existing swipe gesture has continued/changed (finger swiped)

Ended

Swipe gesture has ended (finger lifted)

Cancelled

Swipe was cancelled by user

Happens when interrupted, such as when doing a single-finger swipe and then putting down a second finger.

Failed

Swipe failed... I dunno, ask Apple.

Unknown

Invalid state. Never sent to callbacks.