Module rubrail::util

Utility functions for working with the Apple/TouchBar environment

Contains convenience functions that help with some common actions in Mac environments.

Functions

bundled_resource_path
Locate a resource in the executing Mac App bundle

nsstring_decode
Convert an NSString object into a Rust String

print_nsstring
Print an NSString object to the global logger

set_bg_color⚠
Sets the backgroundColor attribute on an item's view.

set_text_color⚠
Sets the textColor attribute on an item's view.