Function rubrail::util::set_text_color

pub unsafe fn set_text_color(item: &ItemId, r: f64, g: f64, b: f64, alpha: f64)

Sets the textColor attribute on an item's view.

This is an unsafe helper function to set the textColor attribute on the view of an item. It does not verify that the item actually has a view, nor that the view supports textColor, hence unsafe.

Known compatible items: labels

Arguments

item - The ItemId to color

r - Red value (0.0 - 1.0)

g - Green value (0.0 - 1.0)

b - Blue value (0.0 - 1.0)

alpha - Alpha value (0.0 - 1.0)